Separate LinkedIn API helper funcions from share tool
