+++ toolkits/linkedin/arcade_linkedin/tools/constants.py
@@ -1,0 +1,1 @@
+LINKEDIN_BASE_URL = "https://api.linkedin.com/v2"
