+++ toolkits/linkedin/arcade_linkedin/tools/share.py
@@ -1,74 +1,10 @@
 from typing import Annotated
 
-import httpx
 from arcade.sdk import ToolContext, tool
 from arcade.sdk.auth import LinkedIn
 from arcade.sdk.errors import ToolExecutionError
 
-LINKEDIN_BASE_URL = "https://api.linkedin.com/v2"
-
-
-async def _send_linkedin_request(
-    context: ToolContext,
-    method: str,
-    endpoint: str,
-    params: dict | None = None,
-    json_data: dict | None = None,
-) -> httpx.Response:
-    """
-    Send an asynchronous request to the LinkedIn API.
-
-    Args:
-        context: The tool context containing the authorization token.
-        method: The HTTP method (GET, POST, PUT, DELETE, etc.).
-        endpoint: The API endpoint path (e.g., "/ugcPosts").
-        params: Query parameters to include in the request.
-        json_data: JSON data to include in the request body.
-
-    Returns:
-        The response object from the API request.
-
-    Raises:
-        ToolExecutionError: If the request fails for any reason.
-    """
-    url = f"{LINKEDIN_BASE_URL}{endpoint}"
-    token = (
-        context.authorization.token if context.authorization and context.authorization.token else ""
-    )
-    headers = {"Authorization": f"Bearer {token}"}
-
-    async with httpx.AsyncClient() as client:
-        try:
-            response = await client.request(
-                method, url, headers=headers, params=params, json=json_data
-            )
-            response.raise_for_status()
-        except httpx.RequestError as e:
-            raise ToolExecutionError(f"Failed to send request to LinkedIn API: {e}")
-
-    return response
-
-
-def _handle_linkedin_api_error(response: httpx.Response) -> None:
-    """
-    Handle errors from the LinkedIn API by mapping common status codes to ToolExecutionErrors.
-
-    Args:
-        response: The response object from the API request.
-
-    Raises:
-        ToolExecutionError: If the response contains an error status code.
-    """
-    status_code_map = {
-        401: ToolExecutionError("Unauthorized: Invalid or expired token"),
-        403: ToolExecutionError("Forbidden: User does not have Spotify Premium"),
-        429: ToolExecutionError("Too Many Requests: Rate limit exceeded"),
-    }
-
-    if response.status_code in status_code_map:
-        raise status_code_map[response.status_code]
-    elif response.status_code >= 400:
-        raise ToolExecutionError(f"Error: {response.status_code} - {response.text}")
+from arcade_linkedin.tools.utils import _handle_linkedin_api_error, _send_linkedin_request
 
 
 @tool(
@@ -90,6 +26,7 @@ async def create_text_post(
     # LinkedIn calls the user ID "sub" in their user_info data payload. See:
     # https://learn.microsoft.com/en-us/linkedin/consumer/integrations/self-serve/sign-in-with-linkedin-v2#api-request-to-retreive-member-details
     user_id = context.authorization.user_info.get("sub") if context.authorization else None
+
     if not user_id:
         raise ToolExecutionError(
             "User ID not found.",
@@ -110,9 +47,11 @@ async def create_text_post(
     }
 
     response = await _send_linkedin_request(context, "POST", endpoint, json_data=payload)
+
     if response.status_code >= 200 and response.status_code < 300:
         share_id = response.json().get("id")
         return f"https://www.linkedin.com/feed/update/{share_id}/"
 
     _handle_linkedin_api_error(response)
+
     return ""
